More comments and refactoring
- Move filters etc. to their own types, not nested.
- Add more comments.
